+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- JVM desktop effects now run the shared C++ DSP core on the miniaudio callback instead of a Kotlin parameter map
+
 ## [0.2.0] - 2026-08-15
 
 ### Added